Moved AssetsCore and Material into Assets
