+++ Assets/AssetsCore/AssetsManager.h
