+++ Assets/Material/Material.vcxproj
@@ -153,16 +153,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Include\Textures.h" />
-    <ClInclude Include="Include\TexturesInternal.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="Source\Material\TexturesInternal.cpp" />
-  </ItemGroup>
-  <ItemGroup>
     <ProjectReference Include="AssetsCore.vcxproj">
       <Project>{fc37a3e4-4fbe-4088-9de8-2318c80e81e6}</Project>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Textures.h" />
+    <ClInclude Include="TexturesInternal.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="TexturesInternal.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
