+++ Ctrl 2K Engine 0.0.1.sln
@@ -21,11 +21,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Ctrl 2K Renderer", "Ctrl 2K
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Ctrl Core", "Ctrl 2K Engine 0.0.1\Ctrl Core.vcxproj", "{4FC5CA38-4069-48EF-85E2-2ED66EE13F54}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Material", "Ctrl 2K Engine 0.0.1\Material.vcxproj", "{460D5759-6F03-499D-8A8F-DBC87004E0B4}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assets", "Assets", "{26927F71-5167-41ED-8F46-A20AA59F0B40}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AssetsCore", "Ctrl 2K Engine 0.0.1\AssetsCore.vcxproj", "{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}"
 EndProject
@@ -40,6 +36,10 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{3563BD10-00F7-4A33-8B02-4D6FE68EABB8}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial Profiler", "Tools\SunDial Profiler\SunDial Profiler.vcxproj", "{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AssetsCore", "Assets\AssetsCore\AssetsCore.vcxproj", "{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Material", "Assets\Material\Material.vcxproj", "{460D5759-6F03-499D-8A8F-DBC87004E0B4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,22 +81,6 @@ Global
 		{4FC5CA38-4069-48EF-85E2-2ED66EE13F54}.Release|x64.Build.0 = Debug|x64
 		{4FC5CA38-4069-48EF-85E2-2ED66EE13F54}.Release|x86.ActiveCfg = Debug|x64
 		{4FC5CA38-4069-48EF-85E2-2ED66EE13F54}.Release|x86.Build.0 = Debug|x64
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x64.ActiveCfg = Debug|x64
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x64.Build.0 = Debug|x64
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x86.ActiveCfg = Debug|Win32
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x86.Build.0 = Debug|Win32
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x64.ActiveCfg = Release|x64
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x64.Build.0 = Release|x64
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x86.ActiveCfg = Release|Win32
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x86.Build.0 = Release|Win32
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x64.ActiveCfg = Debug|x64
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x64.Build.0 = Debug|x64
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x86.ActiveCfg = Debug|Win32
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x86.Build.0 = Debug|Win32
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x64.ActiveCfg = Release|x64
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x64.Build.0 = Release|x64
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.ActiveCfg = Release|Win32
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.Build.0 = Release|Win32
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.ActiveCfg = Debug|x64
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x64.Build.0 = Debug|x64
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}.Debug|x86.ActiveCfg = Debug|Win32
@@ -137,18 +121,34 @@ Global
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x64.Build.0 = Release|x64
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.ActiveCfg = Release|Win32
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Release|x86.Build.0 = Release|Win32
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x64.ActiveCfg = Debug|x64
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x64.Build.0 = Debug|x64
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x86.ActiveCfg = Debug|Win32
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Debug|x86.Build.0 = Debug|Win32
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x64.ActiveCfg = Release|x64
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x64.Build.0 = Release|x64
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.ActiveCfg = Release|Win32
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}.Release|x86.Build.0 = Release|Win32
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x64.ActiveCfg = Debug|x64
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x64.Build.0 = Debug|x64
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x86.ActiveCfg = Debug|Win32
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Debug|x86.Build.0 = Debug|Win32
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x64.ActiveCfg = Release|x64
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x64.Build.0 = Release|x64
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x86.ActiveCfg = Release|Win32
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{460D5759-6F03-499D-8A8F-DBC87004E0B4} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
-		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{385BC729-58FA-49D3-8609-0A33B280E111} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{55BB3977-E4AA-4C08-B457-B1214977396C} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{EB11A967-DAB8-4DE2-894C-F5014E62137E} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1} = {3563BD10-00F7-4A33-8B02-4D6FE68EABB8}
+		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
+		{460D5759-6F03-499D-8A8F-DBC87004E0B4} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {70BAD174-8BA7-4F92-9815-19B71A819840}
